requirements.txt
APScheduler==3.11.0
python-telegram-bot==22.0
requests==2.32.3
web3==7.11.0
python-dotenv==1.1.0
